requirements.txt
flask
pyjwt
mysql-connector-python